+++ CHANGELOG.md
@@ -2,6 +2,7 @@ master
 
 * Allow specifying multiple file types for checks and syntax check rake files
 * Fix bug where checks which returned lists of lines would output incorrectly
+* Indent check output lines to nest under check name for better readability
 
 0.2.1